Cleaner Code
