+++ src/App.js
@@ -4,6 +4,7 @@ import Menu from './components/MenuComponent';
 import Header from './components/Header';
 import { DISHES } from './shared/dishes';
 class App extends Component {
+  
   constructor(props)
   {
     super(props);
