+++ src/components/Cart/RenderHeading.js
@@ -1,0 +1,19 @@
+import React from 'react';
+
+const RenderHeading = (props) =>
+    {
+        if(props.dishSelected.length>0)
+        {
+            return(
+                <div>
+                <h3>Order Summary</h3>
+                </div>
+            )
+        }
+        else
+        {
+            return null;
+        }
+    }
+
+export default RenderHeading;
